% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{autowd}
\alias{autowd}
\title{Set working dir as autocpx subdir}
\usage{
autowd()
}
\description{
This function allows you to go set the working directory to the "autocpx" subdirectory.
This is necessary to use the document() function in devtools.
}

