% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional prb.R
\name{cprb}
\alias{cprb}
\title{Import probability table.}
\usage{
cprb(cache.value = TRUE)
}
\arguments{
\item{cache.value}{Default TRUE. If you want to pull off of the web again, set to FALSE.}
}
\value{
The name of the data frame generated is \code{prb}. Once the function has completed execution,
the element is available for use.
}
\description{
This function allows the user to pull the table of conditional probabilities into R and perform 
tranformations necessary for further analysis.
}
\details{
The table that is pulled in comes from an online source. Caching is available by default
so the user doesn't have to constantly connect to pull the table in.
}

