% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional variance.R
\name{dat}
\alias{dat}
\title{Data retreival function}
\usage{
dat(offline = FALSE)
}
\arguments{
\item{offline}{Default FALSE. If the file is being obtained locally, then use offline=TRUE to speed loading.
v()}
}
\description{
This function automatically loads data and applies appropriate transformations.
}

