% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional variance.R
\name{gen}
\alias{gen}
\title{Deficiency lookup function.}
\usage{
gen(n, q = 1, offline = FALSE)
}
\arguments{
\item{n}{length of string to generate.}

\item{offline}{Default FALSE. If the file is being obtained locally, then use offline=TRUE to speed loading.
gen()}

\item{k}{number of bits to "look ahead."}
}
\description{
This function generates strings randomly, generates the "look-aheads" and pulls in deficiency data.
}

