% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Deficiency simulator}
\usage{
path(n = 1, file = FALSE, start = 1)
}
\arguments{
\item{n}{Default 1.Number of simulations to generate.}

\item{file}{Default FALSE. If the plot is being printed to file, use file=TRUE.}

\item{start}{Default 1. If you want to start the file numbering at a number other than 1, specify it here.
path()}
}
\description{
This function allows you to randomly generate a 16-length string and plot the deficiency at each point.
}

