% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{upd}
\alias{upd}
\title{Update documentation and reinstall.}
\usage{
upd(OS = "pc")
}
\arguments{
\item{OS}{Enter "pc" if PC and "mac" if mac.}
}
\description{
This function updates the documentation and reinstalls autocpx from local memory.
}

