% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional variance.R
\name{v}
\alias{v}
\title{Name-generating function.}
\usage{
v(k, n = 1, offline = FALSE)
}
\arguments{
\item{k}{Number of bits you want to "look ahead."}

\item{n}{Length of the string (number of iterations.)}

\item{offline}{Default FALSE. If the file is being obtained locally, then use offline=TRUE to speed loading.
v()}
}
\description{
This function allows you to generate names of vectors systematically.
}

