% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{wd}
\alias{wd}
\title{Set working dir}
\usage{
wd(OS = "pc")
}
\arguments{
\item{OS}{Default pc. If on the Mac, use \code{mac} as the argument.}
}
\description{
This function allows you to go straight to the appropriate working directory
when working locally.
}

