% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional variance.R
\name{gen}
\alias{gen}
\title{Deficiency lookup function.}
\usage{
gen(n, q = 1)
}
\arguments{
\item{n}{length of string to generate.}

\item{k}{number of bits to "look ahead."
gen()}
}
\description{
This function generates strings randomly, generates the "look-aheads" and pulls in deficiency data.
}

