% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{v}
\alias{v}
\title{Name-generating function.}
\usage{
v(k, n = 1)
}
\arguments{
\item{k}{Number of bits you want to "look ahead."}

\item{n}{Length of the string (number of iterations.)
v()}
}
\description{
This function allows you to generate names of vectors systematically.
}

